get review return page

+++ src/main/java/phanes/replay/common/dto/response/Page.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class Page<T> {
 
     private Long totalCount;
-    private Integer currentPage;
+    private Long currentPage;
     private T data;
 }
